cloudsql/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asn1crypto==1.5.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloud-sql-python-connector==0.5.2
cryptography==36.0.1
frozenlist==1.3.0
google-api-core==2.7.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
lxml==4.8.0
multidict==6.0.2
numpy==1.22.3
pandas==1.4.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.32
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
